corrected fun fact generator and added to README.  LAST COMMITgit status
