+++ index.html
@@ -7,6 +7,54 @@
     <script src="http://d3js.org/d3.v3.min.js"></script>
     <script src="http://d3js.org/topojson.v1.min.js"></script>
     <script src="js/planetaryjs.min.js"></script>
+    <script type="module">
+        // Import the necessary Firebase modules
+        import { initializeApp } from 'https://www.gstatic.com/firebasejs/9.6.1/firebase-app.js';
+        import { getFirestore, collection, getDocs } from 'https://www.gstatic.com/firebasejs/9.6.1/firebase-firestore.js';
+    
+        // Firebase configuration
+        const firebaseConfig = {
+            apiKey: "AIzaSyCDh4qU36Z0-dkh2ozn9pcP6dxUCT4dZNo",
+            authDomain: "coffeeshoppee-b3a30.firebaseapp.com",
+            databaseURL: "https://coffeeshoppee-b3a30-default-rtdb.firebaseio.com/",
+            projectId: "coffeeshoppee-b3a30",
+            storageBucket: "coffeeshoppee-b3a30.firebasestorage.app",
+            messagingSenderId: "580826414775",
+            appId: "1:580826414775:web:86a502b789b735af05d732"
+        };
+    
+           // Initialize Firebase
+    const app = initializeApp(firebaseConfig);
+    const db = getFirestore(app);
+
+    // Store the fun facts in an array
+    let funFacts = [];
+
+    // Function to fetch fun facts from Firestore
+    async function fetchFunFacts() {
+      const querySnapshot = await getDocs(collection(db, 'fun_facts'));
+      querySnapshot.forEach((doc) => {
+        funFacts.push(doc.data().fact);  // Store each fact in the array
+      });
+// Start displaying fun facts after they're fetched
+      displayFunFact();
+    }
+
+    // Function to display a single fun fact at a time
+    let currentFactIndex = 0;
+    function displayFunFact() {
+      const factElement = document.getElementById('funFact');
+      factElement.textContent = funFacts[currentFactIndex];  // Display the current fun fact
+// Update to the next fact after 5 seconds
+      currentFactIndex = (currentFactIndex + 1) % funFacts.length;  // Loop back to the first fact when done
+    }
+
+    // Fetch and display fun facts when the page loads
+    window.onload = async () => {
+      await fetchFunFacts();
+      setInterval(displayFunFact, 5000);  // Change the fact every 5 seconds
+    };
+        </script>
     <link rel="stylesheet" href="styles.css">
     <script>
         planetaryjs.plugins.objects = function(config) {
@@ -79,28 +127,17 @@
     
 </head>
 <body>
+    
 <center>
 <nav>
     <br>
     <img src="images/Designer (1).png" width="400" alt="CoffeeShoppee Logo" class="logo"><br><br><br>
     <div class="flexcontainer">
         <div class="flexbox">
-            <h1>Random Fun Fact</h1>
-  <div id="fact"></div>
-  <button onclick="fetchFact()">Get New Fact</button>
-
-  <script>
-    async function fetchFact() {
-      try {
-        const response = await fetch("http://localhost:3000/funfact");
-        const data = await response.json();
-        document.getElementById("fact").innerText = data.fact;
-      } catch (error) {
-        console.error("Error fetching fun fact:", error);
-      }
-    }
-  </script>
-
+            <h2>Fun Facts</h2>
+            <ul id="funFact">
+                <!-- Fun facts will be displayed here -->
+              </ul>
             <br>
             <a href="submissionform.html"> Submit Recipe </a>
         </div>
